pyproject.toml
[project]
name = "tishcode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "fastapi>=0.128.0",
    "gitpython>=3.1.46",
    "pygithub>=2.8.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "ruff>=0.14.14",
    "types-requests>=2.32.4.20260107",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
